git-svn-id: http://svn.boost.org/svn/boost/branches/release@63784 b8fc166d-592f-0410-95f2-cb63ce0dd405

+++ libs/serialization/test/test_class_info_save.cpp
@@ -52,7 +52,10 @@ class B
     template<class Archive>
     void serialize(Archive & /*ar*/, const unsigned int file_version){
         // verify at execution that correct version number is passed on save
-        BOOST_CHECK(file_version == ::boost::serialization::version<B>::value);
+        BOOST_CHECK(
+            static_cast<const int>(file_version) 
+            == ::boost::serialization::version<B>::value
+        );
         ++count;
     }
 public:
